Various.

+++ three_little_words/fpmath.h
@@ -7,7 +7,7 @@ typedef int32_t lfp_signed;
 #define FP_UNITY (1 << FP_BITS)
 #define FP_MUL(x, y) (((x) * (y)) >> FP_BITS)
 #define FP_DIV(n, d) (((n)<<FP_BITS) / (d))
-#define FP2FLOAT(x) ((x) / ((float)(FP_UNITY)))
+#define FP2FLOAT(x) ((x) / ((double)(FP_UNITY)))
 #define FLOAT2FP(x) ((fp_signed)((x) * FP_UNITY))
 #define FLOAT2LFP(x) ((lfp_signed)((x) * FP_UNITY))
